update elements frame membership after align/distribute action
